Fix inverted test names
